stage 291019-1202

+++ app/Entity/Shop/Catalog/Brand/BrandObserver.php
@@ -46,7 +46,9 @@ class BrandObserver
      */
     public function deleted(Brand $brand)
     {
-        $this->removeFile($brand->image);
+        if ($brand->image) {
+            $this->removeFile($brand->image);
+        }
     }
 
     /**
